Refactoring common code between the PT and TF trainer into an util file.
